Finish exercise 2 of odin's flex exercises
As with past others, everything good; the only thing
I'll say I did different was that I inlined the list items with
display inline, instead of making the ul a flex container. It did not
said I could not use inline so I passed the self check. Still
the other solution was better.